producer/requirements.txt
fastapi
uvicorn[standard]
pika
python-dotenv
pydantic
requests